tiles_management/tiles_management/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from inventory.views import (
    BrandViewSet, TileCategoryViewSet, LocationViewSet,
    ProductViewSet, InventoryViewSet, StockMovementViewSet
)

router = DefaultRouter()
router.register(r'brands', BrandViewSet)
router.register(r'categories', TileCategoryViewSet)
router.register(r'locations', LocationViewSet)
router.register(r'products', ProductViewSet)
router.register(r'inventory', InventoryViewSet)
router.register(r'stock-movements', StockMovementViewSet)

urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include(router.urls)),
]

CSRF_TRUSTED_ORIGINS = ['https://*.vercel.app']